requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
selenium>=4.8.0
webdriver-manager>=4.8.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.30.0